+++ pysbolgraph/utils/S2Cut.py
@@ -1,6 +1,6 @@
-from .S2Location import S2Location
+from S2Location import S2Location
 
-from .terms import SBOL2
+from meta.terms import SBOL2
 
 
 class S2Cut(S2Location):
