+++ pysbolgraph/utils/S2Sequence.py
@@ -1,6 +1,6 @@
-from .S2Identified import S2Identified
+from S2Identified import S2Identified
 
-from .terms import SBOL2
+from meta.terms import SBOL2
 
 
 class S2Sequence(S2Identified):
