+++ pysbolgraph/utils/meta/CompliantURIs.py
